test: move test components into individuals files
Inspired by a comment from @mbayopanda, this commit moves test harnesses
into their own files to prevent git conflicts in the future.  The method
of using these test harnesses has not been changed.
